Move ExecutorEventListener to core
